16052025 experiment for enabling cred in the jem 2100f in darmstadt trough goniotool.
it commit many files for different line separator, but should not bothering me. to check.

changes:
goniotool_socket.py (now we are just using the wrapper class goniotool without server client interaction! goniotool must be open in the same pc as pyfastadt, recommended camera pc, it works smoothly)
adaptor_jeol.py (now jeol adaptor ask espicitely for goniotool in initialization! it must be opened otherwise cred doesn't work. set_alpha have the optional wait that is used to recognize if continuous rotation is used. in that case the gonio rotate only if the main thread of the camera launch the rotation after you press yes. if wait == False, it rotate the stage for you as usual. implemented method continuous rotation to enable cred in jeol trough goniotool. tested it works at the speed close to the one that you choose. we recognized also that a thread like stage_thread is not able to launch simple methods from the microscope this can be an issue for the future, and a server client setup should be durable to enable beam and stage movements in parallel. otherwise another thing to check is to pass to multithreaded environment using comtypes maybe this is the most elegant way to do that. )
jem2100f_speed_lookuptable.csv (the lookup table built from excel contain tabs that must be respected otherwise pandas complains for the presence of NaN or in general string to float operation not allowed when try to understand the correct speed from the tabulated values!)
merlin_table.txt (added some calibration and the correct ipadress of goniotool inside)
adaptor_merlin.py (in the methodacquire_cred_data after the set event to start the acquisition now the camera send also the signal to start the alpha tilt, passing trough self.instance_gui.tem.set_alpha. fixed the timeout for cred in merlin we were not sending the SOFTTRIGGER command and so nothing was there to be retrieved. the acquisition of the images is still based on the timing of the stage_thread that is monitoring the angles to arrive to the final angle. it works. )

todo in future bugfixing: i noticed that the calculation for the buffer size is wrong, the system allocate not enough frames, like half of the required!! to check it it's probably in the function retrieve_parameters_for_acquisition in fast_adt_func.py

apart from this problem cred should colelct data right now at the right speed. to test if it's correct with a real sample.